+++ app/[locale]/login/page.tsx
@@ -1,25 +1,25 @@
 'use client'
 
-import React, { useEffect, useRef, useState } from "react";
-import { useRouter } from 'next/navigation';
+import React, {useEffect, useRef, useState} from "react";
+import {useRouter} from 'next/navigation';
 import logECS from '@/utils/clientLogger';
-import { useCsrf } from '@/context/CsrfContext';
-import { handleErrors } from '@/utils/errorHandler';
-import { useAuthContext } from '@/context/AuthContext';
-import { useTranslations } from "next-intl";
+import {useCsrf} from '@/context/CsrfContext';
+import {handleErrors} from '@/utils/errorHandler';
+import {useAuthContext} from '@/context/AuthContext';
+import {useTranslations} from "next-intl";
 import {
   Button,
+  FieldError,
   Form,
-  TextField,
+  Input,
+  Label,
   Link,
   Text,
-  Label,
-  Input,
-  FieldError,
+  TextField,
 } from "react-aria-components";
 import {
-  LayoutContainer,
   ContentContainer,
+  LayoutContainer,
   ToolbarContainer,
 } from '@/components/Container';
 import ErrorMessages from "@/components/ErrorMessages";