+++ components/SelectExistingTemplate/index.tsx
@@ -17,10 +17,7 @@ import {
 
 //Components
 import PageHeader from "@/components/PageHeader";
-import {
-  ContentContainer,
-  LayoutContainer,
-} from '@/components/Container';
+import { ContentContainer, LayoutContainer, } from '@/components/Container';
 import { filterTemplates } from '@/components/SelectExistingTemplate/utils';
 import TemplateList from '@/components/TemplateList';
 import ErrorMessages from '@/components/ErrorMessages';
@@ -28,8 +25,8 @@ import ErrorMessages from '@/components/ErrorMessages';
 //GraphQL
 import {
   useAddTemplateMutation,
-  usePublishedTemplatesQuery,
-  useMyVersionedTemplatesQuery
+  useMyVersionedTemplatesQuery,
+  usePublishedTemplatesQuery
 } from '@/generated/graphql';
 
 // Hooks